yzksp-app/src/backend/yzkspProject/requirements.txt
Django>=3.0,<4.0
djangorestframework
django-cors-headers
django-allauth